creating ennums module to avoid circular imports issues
