+++ packages/pymodaq/src/pymodaq/control_modules/enums.py
@@ -1,0 +1,38 @@
+from pymodaq_gui.plotting.data_viewers import ViewersEnum
+from pymodaq_utils.enums import BaseEnum
+
+
+class DAQTypesEnum(BaseEnum):
+    """enum relating a given DAQType and a viewer type
+    See Also
+    --------
+    pymodaq.utils.plotting.data_viewers.viewer.ViewersEnum
+    """
+    DAQ0D = 'Viewer0D'
+    DAQ1D = 'Viewer1D'
+    DAQ2D = 'Viewer2D'
+    DAQND = 'ViewerND'
+
+    def to_data_type(self):
+        return ViewersEnum[self.value].value
+
+    def to_viewer_type(self):
+        return self.value
+
+    def to_daq_type(self):
+        return self.name
+
+    def increase_dim(self, ndim: int):
+        dim = self.get_dim()
+        if dim != 'N':
+            dim_as_int = int(dim) + ndim
+            if dim_as_int > 2:
+                dim = 'N'
+            else:
+                dim = str(dim_as_int)
+        else:
+            dim = 'N'
+        return DAQTypesEnum(f'Viewer{dim}D')
+
+    def get_dim(self):
+        return self.value.split('Viewer')[1].split('D')[0]
