+++ packages/pymodaq/src/pymodaq/control_modules/instruments.py
@@ -1,12 +1,9 @@
 from typing import TYPE_CHECKING, Type
 from pymodaq_gui.parameter import Parameter
 
-from pymodaq_gui.plotting.data_viewers import ViewersEnum
-from pymodaq_utils.enums import BaseEnum
-
 from pymodaq_utils.utils import find_dict_in_list_from_key_val, find_dicts_in_list_from_key_val
 from pymodaq.utils.exceptions import DetectorError, ActuatorError
-
+from .enums import DAQTypesEnum  # noqa for backcompatibility
 from pymodaq import CONTROL_MODULES
 
 
@@ -36,41 +33,6 @@ def find_actuator_class_from_name(actuator_name: str) -> Type['DAQ_Move_base']:
         "DAQ_Move_" + actuator_name,
     )
     return class_
-
-class DAQTypesEnum(BaseEnum):
-    """enum relating a given DAQType and a viewer type
-    See Also
-    --------
-    pymodaq.utils.plotting.data_viewers.viewer.ViewersEnum
-    """
-    DAQ0D = 'Viewer0D'
-    DAQ1D = 'Viewer1D'
-    DAQ2D = 'Viewer2D'
-    DAQND = 'ViewerND'
-
-    def to_data_type(self):
-        return ViewersEnum[self.value].value
-
-    def to_viewer_type(self):
-        return self.value
-
-    def to_daq_type(self):
-        return self.name
-
-    def increase_dim(self, ndim: int):
-        dim = self.get_dim()
-        if dim != 'N':
-            dim_as_int = int(dim) + ndim
-            if dim_as_int > 2:
-                dim = 'N'
-            else:
-                dim = str(dim_as_int)
-        else:
-            dim = 'N'
-        return DAQTypesEnum(f'Viewer{dim}D')
-
-    def get_dim(self):
-        return self.value.split('Viewer')[1].split('D')[0]
 
 
 def get_viewer_plugins(daq_type, det_name):
