+++ packages/pymodaq/tests/control_modules/daq_viewer_test.py
@@ -14,7 +14,8 @@ from pymodaq.control_modules import daq_viewer as daqvm
 from pymodaq.control_modules.daq_viewer import DAQ_Viewer
 from pymodaq.control_modules.viewer_utility_classes import HW_SETTINGS_KEY as DETECTOR_SETTINGS_KEY
 from pymodaq.control_modules.utils import ControlModule
-from pymodaq.control_modules.instruments import DET_TYPES, get_viewer_plugins, DAQTypesEnum
+from pymodaq.control_modules.instruments import DET_TYPES, get_viewer_plugins
+from pymodaq.control_modules.enums import DAQTypesEnum
 from pymodaq.utils.conftests import qtbotskip, main_modules_skip
 from pymodaq.utils.config import GlobalConfig
